fix(types): read01 round 189 — scalar subquery results keep their type (count is bigint)
The r188 pin caught it: every scalar-subquery shape typed count(*)
as `integer` (PG: `bigint`) — a typed-decode breaker for sqlx
clients (the mailrs round-27 class of bug). Three narrowing sites:

1. value_to_literal_expr materialised BigInt/SmallInt results as
   bare Literal::Integer, which re-types to INT when the value fits
   i32; now wrapped in explicit casts (::int8 / ::int2) so the
   re-evaluated literal keeps the type (both strict and permissive
   variants).
2. scalar_subquery_empty_default returned Int(0) for the empty-set
   count — now BigInt(0).
3. The SCALARSQ PK-probe fast paths (probe_with_pk_fast_path + the
   row-eval twin + the plan placeholder) returned Int(0)/Int(1) —
   now BigInt, matching the slow path they replace.

Differential: 5/5 pg_typeof probes SAME after the fix (count/sum
bigint on plain + correlated + VALUES shapes, ::smallint preserved,
arithmetic values exact). One legacy embedded pin had encoded the
narrowed Int as expected ("COUNT(*) surfaces as Int on this path")
— corrected to BigInt. r188's value pin restored to its precise
form. Engine + embedded suites green x3 consecutive runs, clippy
clean.

+++ crates/spg-embedded/tests/e2e/mailrs_round15_16.rs
@@ -120,9 +120,11 @@ fn correlated_not_exists_on_readonly_and_join_paths() {
          WHERE ac.message_id = m.id) FROM messages m ORDER BY m.id",
     );
     assert_eq!(r.len(), 3);
-    // COUNT(*) in a scalar subquery surfaces as Int on this path.
-    assert_eq!(r[0][1], Value::Int(1));
-    assert_eq!(r[1][1], Value::Int(0));
+    // v7.39 (round 189) — COUNT(*) in a scalar subquery is BIGINT
+    // (PG's type); the old Int here pinned the literal-round-trip
+    // narrowing this round fixed.
+    assert_eq!(r[0][1], Value::BigInt(1));
+    assert_eq!(r[1][1], Value::BigInt(0));
 }
 
 /// Round-16 collateral #1 — multi-row VALUES drew the SAME serial id
